+++ pyrio/__init__.py
@@ -1,3 +1,4 @@
-from .stream import Stream as Stream
-from .file_stream import FileStream as FileStream
-from .optional import Optional as Optional
+from pyrio.streams.stream import Stream as Stream
+from pyrio.streams.file_stream import FileStream as FileStream
+from pyrio.utils.optional import Optional as Optional
+from pyrio.utils.dict_item import Item as Item
