+++ pyrio/utils/exception.py
